Move `data-types` into `web-common/components`
