+++ web-common/src/components/data-types/DataTypeIcon.svelte
