+++ web-common/src/components/data-types/FormattedDataType.svelte
