+++ web-common/src/components/data-types/Interval.svelte
