+++ web-common/src/components/data-types/Number.svelte
