+++ web-local/src/lib/components/column-profile/ProfileContainer.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import { FormattedDataType } from "@rilldata/web-common/components/data-types";
   import { createShiftClickAction } from "@rilldata/web-local/lib/util/shift-click-action";
   import { createEventDispatcher } from "svelte";
   import { slide } from "svelte/transition";
@@ -6,7 +7,6 @@
     COLUMN_PROFILE_CONFIG,
     LIST_SLIDE_DURATION,
   } from "../../application-config";
-  import { FormattedDataType } from "../data-types";
 
   const dispatch = createEventDispatcher();
   const { shiftClickAction } = createShiftClickAction();
