+++ web-local/src/lib/components/column-profile/column-types/NumericProfile.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
+  import { DataTypeIcon } from "@rilldata/web-common/components/data-types";
   import { useRuntimeServiceGetDescriptiveStatistics } from "@rilldata/web-common/runtime-client";
   import { httpRequestQueue } from "@rilldata/web-common/runtime-client/http-client";
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
   import { getPriorityForColumn } from "@rilldata/web-local/lib/http-request-queue/priorities";
   import { copyToClipboard } from "@rilldata/web-local/lib/util/shift-click-action";
   import { derived } from "svelte/store";
-  import { DataTypeIcon } from "../../data-types";
   import ProfileContainer from "../ProfileContainer.svelte";
   import {
     getNullPercentage,
