+++ web-local/src/lib/components/column-profile/column-types/TimestampProfile.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
+  import { DataTypeIcon } from "@rilldata/web-common/components/data-types";
+  import Interval from "@rilldata/web-common/components/data-types/Interval.svelte";
   import { TIMESTAMP_TOKENS } from "@rilldata/web-common/lib/duckdb-data-types";
   import { httpRequestQueue } from "@rilldata/web-common/runtime-client/http-client";
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
-  import { copyToClipboard } from "@rilldata/web-local/lib/util/shift-click-action";
+  import { copyToClipboard } from "../../../util/shift-click-action";
   import TimestampDetail from "../../data-graphic/compositions/timestamp-profile/TimestampDetail.svelte";
   import TimestampSpark from "../../data-graphic/compositions/timestamp-profile/TimestampSpark.svelte";
   import WithParentClientRect from "../../data-graphic/functional-components/WithParentClientRect.svelte";
-  import { DataTypeIcon } from "../../data-types";
-  import Interval from "../../data-types/Interval.svelte";
   import ProfileContainer from "../ProfileContainer.svelte";
   import { getNullPercentage, getTimeSeriesAndSpark } from "../queries";
   import NullPercentageSpark from "./sparks/NullPercentageSpark.svelte";
