+++ web-local/src/lib/components/column-profile/column-types/VarcharProfile.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
+  import { DataTypeIcon } from "@rilldata/web-common/components/data-types";
   import { httpRequestQueue } from "@rilldata/web-common/runtime-client/http-client";
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
   import { copyToClipboard } from "@rilldata/web-local/lib/util/shift-click-action";
-  import { DataTypeIcon } from "../../data-types";
   import ProfileContainer from "../ProfileContainer.svelte";
   import { getCountDistinct, getNullPercentage, getTopK } from "../queries";
   import TopK from "./details/TopK.svelte";
